chore: refactor the rfi upload common code
